feat: add playGame()
playGame() lets you play up to 5 rounds of the game.
This is basically the UI of the game.

Therefore, I refactored playRound() by moving the
scoring to playGame(), because playRound() is solely
for the logic of the game.

Also, I added confirm() in playGame(), so that
the user can exit the game whenever he/she wants.